+++ backend/src/main/java/com/example/demo/repository/PdfRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.entity.Pdf;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PdfRepository extends JpaRepository<Pdf, Long> {
+public interface PdfRepository extends JpaRepository<Pdf, Integer> {
 }
